requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.2
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.79.0
openpyxl>=3.1.1
xlrd>=2.0.1
xlsxwriter>=3.0.0